Implement Animation::setTimeline
Mutability of the animation timeline is behind the feature flag for
ScrollTimeline. The feature is marked as readonly in the auto-generated
web-animations IDL in WPT since marked as readonly in web-animations-1.
This discrepancy introduces one IDL harness failure with the change.
The feature is marked as web-exposed, but with a no-op implementation
if the feature flag is not set.

With the patch, a number of the failing tests are now failing for
different reasons:
* Replaceable animations should consider all timelines when determining
  if an animation can be removed. Our current implementation only
  considers other animations on the same timeline when making the
  decision. Fixing these failures requires some refactoring, which is
  deferred to a followup CL.
* Resetting the timeline cancels a running CSS animation or transition.
  Some refactoring is required to ensure that we properly compute the
  cancel time as it depends on the old timeline. Again, deferring to
  a followup CL.

https://github.com/w3c/csswg-drafts/issues/5159
https://drafts.csswg.org/scroll-animations-1
https://www.chromestatus.com/feature/6752840701706240

Bug: 827626
Change-Id: I1fea09a80e8730694455c06b192cb9e068eb0c55